Update style to better highlight selected tab
With the current steup there is no possibility to have the selected tab
not have a bottom border, while the regular tabs.

This commit fills the selected tab with the border color, which is better
the other way around.

This is also consistent with other selected backgound items in the
Projects and Open Documents views.